requirements.txt
asgiref                  3.4.1
dj-database-url          0.5.0
Django                   3.2.9
django-cors-headers      3.10.0
django-filter            21.1
django-heroku            0.3.1
django-phone-field       1.8.1
django-phonenumber-field 5.2.0
djangorestframework      3.12.4
gunicorn                 20.1.0
Markdown                 3.3.4
phonenumbers             8.12.36
pip                      21.3.1
psycopg2                 2.9.1
pytz                     2021.3
setuptools               56.0.0
sqlparse                 0.4.2
whitenoise               5.3.0